Some Change And Fix.
- Make tackle distance longer.
- Make spin down longer than spin up.
- Adjust sprite captain-spin and sprite ElectricTrap.
- Fix weird sound when change scene with shortcut checkpoint shortcut while being in ambush.
- Move the try catch block from main function to the game loop.
- Adjust MovingLedge speed.
- Fix EnemyRocketBackAndForth to not shoot right away, fix EnemyRocketBackAndForthJumping.
- Change Captain max-height-jump to be lower if jump is from water.
- Fix Captain being stuck in Injured state.
- Adjust bunker-bullet speed.
- Fix ElectricBat weird spawn.
- Make ElectricBat not to explode when hit Captain while flyattacking.
- Fix weird mixed sound when Captain dead.
- Adjust Captain walking speed.
- Make handling shield collision more precise.

+++ SourceProject/BulletEnemyWizard.cpp
@@ -49,6 +49,6 @@ RectF BulletEnemyWizard::GetBBox() const
 void BulletEnemyWizard::Reflect()
 {
 	isReflected = true;
-	vel.x = -nx * BULLET_MOVING/sqrt(2);
+	vel.x = -nx * BULLET_MOVING/sqrtf(2.0f);
 	vel.y = -100;
 }
